cli/requirements.txt
typer~=0.12.0
rich~=13.7.0
PyYAML==6.0.1
requests>=2.26.0
pydantic==1.10.2
yaspin==2.1.0
tabulate==0.9.0
pexpect==4.8.0
colorama==0.4.4
time-machine==2.4.0
pytest-mock==3.14.0
pyfakefs==5.8.0
validators==0.18.2
merge-args==0.1.4
synapseclient==4.1.1
schema==0.7.5
email-validator==2.0.0
auth0-python==4.3.0
pandas==2.1.0
numpy==1.26.4
watchdog==3.0.0
GitPython==3.1.41
psutil==5.9.8
semver==3.0.4
cookiecutter==2.1.1
uvicorn==0.30.1
fastapi==0.111.1
fastapi-login==1.10.2
cwltool==3.1.20250110105449
cwl-refrunner=1.0